requirements.txt
python-decouple
requests
psutil